pyproject.toml
[project]
name = "usgs-chart-rag"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "qdrant-client>=1.15.0,<1.16",
    "openai>=1.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "tqdm>=4.66.0",
    "langchain>=0.1.0,<0.3.0",
    "langchain-openai>=0.1.0",
    "playwright>=1.42.0,<2.0.0",
    "numpy>=1.23.0",
    "langsmith>=0.1.147",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    # later you can add: "pytest-cov", "ruff", etc.
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
